+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All Notable changes to `enqueuer` will be documented in this file
 ## 1.1.0 - 2015-12-13
 
 ### Added
-- More public options for clearing cache
+- More public options for clearing cache including magic methods
 - Added the option to include a data object with a script to pass along php variables to scripts
 
 ## 1.0.0 - 2015-12-12